添加了Filter过滤器，并且通过Configuration配置到Spring容器中，实现了RedisOperate注解AOP解释，添加了分类分页缓存，以及更新删除时删除所有reids缓存。

+++ CangQiong/src/main/java/com/example/cangqiong/Mapper/EmployeeMapper.java
@@ -36,11 +36,11 @@ public interface EmployeeMapper {
 
     //修改密码
     @Update("update employee set password = #{password} where id = #{id}")
-    Integer updatePassword(@Param("password") String password , @Param("id") Integer id);
+    Integer updatePassword(@Param("password") String password , @Param("id") Long id);
 
     //得到原始密码
     @Select("select password from employee where id = #{id}")
-    String selectOldPassword(Integer id);
+    String selectOldPassword(Long id);
 
     //更改员工状态
     @Update("update employee set status = #{status} where id = #{id}")
